+++ crates/polars-arrow/src/lib.rs
@@ -37,6 +37,7 @@ pub mod datatypes;
 pub mod ffi;
 pub mod legacy;
 pub mod util;
+pub mod pushable;
 
 // re-exported because we return `Either` in our public API
 // re-exported to construct dictionaries
